Add files via upload
Date: 19/03/2019

Problem Statement: Good morning.
This is your coding interview problem for today.
This problem was recently asked by Google.
Given a list of numbers and a number k, return whether any two numbers from the list add up to k.
For example, given [10, 15, 3, 7] and k of 17, return true since 10 + 7 is 17.
Bonus: Can you do this in one pass?

Additional Comments: Solution erroneously handles the case where a pair is not found. However, the fix is too easy to be bothered with.